Turn `eval.py` into package
