+++ arabot/cogs/eval/abc.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+from typing import Any, Protocol
+
+
+class Evaluator(Protocol):
+    TB_FILENAME = "<repl>"
+
+    @abstractmethod
+    async def run(self, code: str, *args, **kwargs) -> tuple[str, Any]:
+        ...
